gemini_integration.py
"""
Módulo para integração futura com a API do Google Gemini
"""
import os
from typing import Optional, Dict, Any
import google.generativeai as genai
from datetime import datetime

class GeminiIntegration:
    """
    Classe para integração com a API do Google Gemini
    """
    
    def __init__(self, api_key: Optional[str] = None):
        self.api_key = api_key or os.getenv('GEMINI_API_KEY')
        self.model = None
        self.chat_session = None
        
        if self.api_key:
            self.initialize_gemini()
    
    def initialize_gemini(self):
        """Inicializa a conexão com a API do Gemini"""
        try:
            genai.configure(api_key=self.api_key)
            self.model = genai.GenerativeModel('gemini-pro')
            print("✅ Gemini API inicializada com sucesso!")
        except Exception as e:
            print(f"❌ Erro ao inicializar Gemini API: {e}")
            self.model = None
    
    def generate_response(self, message: str, context: Optional[str] = None) -> Dict[str, Any]:
        """
        Gera resposta usando a API do Gemini
        
        Args:
            message: Mensagem do usuário
            context: Contexto adicional da conversa
            
        Returns:
            Dict com resposta e metadados
        """
        if not self.model:
            return {
                'response': 'Desculpe, a integração com Gemini não está disponível no momento.',
                'error': 'Gemini API não inicializada',
                'timestamp': datetime.now().isoformat()
            }
        
        try:
            # Preparar prompt com contexto se disponível
            prompt = message
            if context:
                prompt = f"Contexto da conversa: {context}\n\nMensagem do usuário: {message}"
            
            # Gerar resposta
            response = self.model.generate_content(prompt)
            
            return {
                'response': response.text,
                'timestamp': datetime.now().isoformat(),
                'model': 'gemini-pro',
                'success': True
            }
            
        except Exception as e:
            return {
                'response': f'Desculpe, ocorreu um erro ao processar sua mensagem com o Gemini: {str(e)}',
                'error': str(e),
                'timestamp': datetime.now().isoformat(),
                'success': False
            }
    
    def start_chat_session(self, system_prompt: Optional[str] = None):
        """
        Inicia uma nova sessão de chat com o Gemini
        
        Args:
            system_prompt: Prompt do sistema para definir comportamento
        """
        if not self.model:
            return False
        
        try:
            if system_prompt:
                self.chat_session = self.model.start_chat(history=[])
                # Enviar prompt do sistema como primeira mensagem
                self.chat_session.send_message(system_prompt)
            else:
                self.chat_session = self.model.start_chat(history=[])
            
            return True
        except Exception as e:
            print(f"Erro ao iniciar sessão de chat: {e}")
            return False
    
    def send_message_to_session(self, message: str) -> Dict[str, Any]:
        """
        Envia mensagem para a sessão de chat ativa
        
        Args:
            message: Mensagem do usuário
            
        Returns:
            Dict com resposta e metadados
        """
        if not self.chat_session:
            return self.generate_response(message)
        
        try:
            response = self.chat_session.send_message(message)
            
            return {
                'response': response.text,
                'timestamp': datetime.now().isoformat(),
                'model': 'gemini-pro',
                'success': True,
                'session_active': True
            }
            
        except Exception as e:
            return {
                'response': f'Erro na sessão de chat: {str(e)}',
                'error': str(e),
                'timestamp': datetime.now().isoformat(),
                'success': False
            }
    
    def get_chat_history(self) -> list:
        """Obtém histórico da sessão de chat atual"""
        if not self.chat_session:
            return []
        
        try:
            return self.chat_session.history
        except Exception as e:
            print(f"Erro ao obter histórico: {e}")
            return []
    
    def clear_chat_session(self):
        """Limpa a sessão de chat atual"""
        self.chat_session = None
    
    def is_available(self) -> bool:
        """Verifica se a integração com Gemini está disponível"""
        return self.model is not None and self.api_key is not None
    
    def get_model_info(self) -> Dict[str, Any]:
        """Obtém informações sobre o modelo Gemini"""
        if not self.model:
            return {'available': False}
        
        return {
            'available': True,
            'model_name': 'gemini-pro',
            'api_key_set': bool(self.api_key),
            'chat_session_active': self.chat_session is not None
        }

# Função para configurar Gemini no chatbot principal
def setup_gemini_in_chatbot(chatbot_instance, api_key: Optional[str] = None):
    """
    Configura a integração Gemini no chatbot principal
    
    Args:
        chatbot_instance: Instância do ChatBot
        api_key: Chave da API do Gemini (opcional)
    """
    try:
        gemini = GeminiIntegration(api_key)
        if gemini.is_available():
            chatbot_instance.set_gemini_integration(gemini)
            print("✅ Integração Gemini configurada com sucesso!")
            return True
        else:
            print("⚠️ Gemini não disponível - usando respostas padrão")
            return False
    except Exception as e:
        print(f"❌ Erro ao configurar Gemini: {e}")
        return False

# Exemplo de uso
if __name__ == "__main__":
    # Teste da integração
    gemini = GeminiIntegration()
    
    if gemini.is_available():
        print("Testando integração com Gemini...")
        response = gemini.generate_response("Olá! Como você funciona?")
        print(f"Resposta: {response['response']}")
    else:
        print("Gemini não disponível. Configure a variável GEMINI_API_KEY.")
