requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
python-dotenv==1.0.0

# Para integração futura com Gemini
google-generativeai==0.3.2
requests==2.31.0

# Para desenvolvimento
python-dotenv==1.0.0
